requirements.txt
azure-functions
dnspython
requests
